moving networking all together
